Add FailHandler for node.
FailHandler is an optional function that will be called to handle an
IOnDiskStateMachine failure. If set, the replica will be stopped when state
machine returns an error (from all methods except Lookup) afterwards the
FailHandler will be called.  This is intended only for cases when
statemachine cannot continue i.e. disk failure but you do not want
dragonboat to panic.